Condicional de busca para leitor
feita condicional para ativar leitor somente quando vier link novo e condicional no leitor para baixar e ler somente links novos no banco de dados

+++ backend/src/robo_leitor_pdf/leitor.py
@@ -6,15 +6,15 @@ from PyPDF2 import PdfFileReader
 import mysql.connector
 
 def leitor():
-    db_connection = mysql.connector.connect(host="localhost", user="root", passwd="root", database="apimidall")
+    db_connection = mysql.connector.connect(host="localhost", user="MatheusAJesus", passwd="@Fatec2022", database="apimidall")
     cursor = db_connection.cursor(buffered=True)
 
 
     def select(sql):
         cursor.execute(sql)
         listaDaConsultaDosProcessos = []
-        for (idAssociado, nome, id_processo, link ) in cursor:
-            listaDaConsultaDosProcessos.append([idAssociado, nome, id_processo, link])
+        for (idAssociado, nome, id_processo, conteudo, link ) in cursor:
+            listaDaConsultaDosProcessos.append([idAssociado, nome, id_processo, conteudo, link])
         return listaDaConsultaDosProcessos
 
     def update(sql, valor, valor2):
@@ -22,23 +22,24 @@ def leitor():
         valor2 = valor2
         return cursor.execute(sql,(valor,valor2))
             
-    consulta = select("select a.id, a.nome, p.id, p.link  from associados a, processos p where a.id = p.associadoId;")
+    consulta = select("select a.id, a.nome, p.id, p.conteudo, p.link  from associados a, processos p where a.id = p.associadoId;")
     for consultaPosicao in consulta:
-        wget.download(consultaPosicao[3], f'PDFs/{consultaPosicao[1]}.pdf')
-        pdf_para_leitura = PdfFileReader(f'PDFs/{consultaPosicao[1]}.pdf')
-        page_object = pdf_para_leitura.getPage(0)
-        texto_do_pdf = page_object.extract_text()
-        print(page_object)
-        blocos_de_texto = texto_do_pdf.split('\n')
-        for i in blocos_de_texto:
-            nome= consultaPosicao[1]
-            if re.search(nome, i, re.IGNORECASE):
-                bloco_relacionado_associado = i
-                bloco_relacionado_associado = '...'+bloco_relacionado_associado+'...'
-                sql = ("update processos set conteudo = %s"
-                        "where id= %s")
-                up = update(sql, bloco_relacionado_associado, consultaPosicao[2])
-                os.remove(f'PDFs/{consultaPosicao[1]}.pdf')
+        if consultaPosicao[3] == None:
+            wget.download(consultaPosicao[4], f'PDFs/{consultaPosicao[1]}.pdf')
+            pdf_para_leitura = PdfFileReader(f'PDFs/{consultaPosicao[1]}.pdf')
+            page_object = pdf_para_leitura.getPage(0)
+            texto_do_pdf = page_object.extract_text()
+            print(page_object)
+            blocos_de_texto = texto_do_pdf.split('\n')
+            for i in blocos_de_texto:
+                nome= consultaPosicao[1]
+                if re.search(nome, i, re.IGNORECASE):
+                    bloco_relacionado_associado = i
+                    bloco_relacionado_associado = '...'+bloco_relacionado_associado+'...'
+                    sql = ("update processos set conteudo = %s"
+                            "where id= %s")
+                    update(sql, bloco_relacionado_associado, consultaPosicao[2])
+                    os.remove(f'PDFs/{consultaPosicao[1]}.pdf')
 
     cursor.close()
     db_connection.commit()
